Add +k8s:deepcopy-gen:interfaces tags to runtime.Object impls in tests
