+++ pkg/cmd/util/pluginconfig/testing/doc.go
@@ -1,0 +1,4 @@
+// +k8s:deepcopy-gen=package
+
+// Package testing contains types for tests.
+package testing
