+++ pkg/util/testing/doc.go
@@ -1,0 +1,4 @@
+// +k8s:deepcopy-gen=package
+
+// Package testing contains testing types.
+package testing
